+++ src/Contracts/NameOrder.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 
 namespace Konekt\Address\Contracts;
 
-interface NameOrder
+use Konekt\Enum\EnumInterface;
+
+interface NameOrder extends EnumInterface
 {
 }