+++ src/resources/database/migrations/2023_06_15_174343_add_model_to_the_addresses_table.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class () extends Migration {
+    public function up(): void
+    {
+        Schema::table('addresses', function (Blueprint $table) {
+            $table->nullableMorphs('model');
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::table('addresses', function (Blueprint $table) {
+            $table->dropMorphs('model');
+        });
+    }
+};