+++ Changelog.md
@@ -9,7 +9,9 @@
 - Changed minimum Enum version to v4.1
 - BC: Changed all enum interfaces to extend the root `EnumInterface`
 - BC: added the `iso2Code()` and the `getName()` methods to the `Country` interface
+- BC: added the `model()` method to the `Address` interface
 - Added firstname, lastname, company_name, email, phone, address2, access_code, tax_nr and registration_nr fields to the Address table/model
+- Added a generic, optional and polymorphic `model` relationship to the Address model
 
 ## 2.8.0
 ##### 2023-04-08